refactor: use separate files for ept and copc loaders and geometry
