+++ src/PotreeRenderer.js
@@ -281,7 +281,7 @@ class Shader {
 			}
 
 			// uniform blocks
-			if(gl instanceof WebGL2RenderingContext){ 
+			if(gl instanceof WebGL2RenderingContext){
 				let numBlocks = gl.getProgramParameter(program, gl.ACTIVE_UNIFORM_BLOCKS);
 
 				for (let i = 0; i < numBlocks; i++) {
@@ -294,7 +294,7 @@ class Shader {
 					gl.uniformBlockBinding(program, blockIndex, blockIndex);
 					let dataSize = gl.getActiveUniformBlockParameter(program, blockIndex, gl.UNIFORM_BLOCK_DATA_SIZE);
 
-					let uBuffer = gl.createBuffer();	
+					let uBuffer = gl.createBuffer();
 					gl.bindBuffer(gl.UNIFORM_BUFFER, uBuffer);
 					gl.bufferData(gl.UNIFORM_BUFFER, dataSize, gl.DYNAMIC_READ);
 
@@ -574,6 +574,13 @@ export class Renderer {
 		let gl = this.gl;
 		let webglBuffer = new WebGLBuffer();
 		webglBuffer.vao = gl.createVertexArray();
+
+		// debugger
+		// console.group("Renderer.createBuffer")
+		// console.log("geometry...", geometry)
+		// console.log("geometry.attributes.position.count...", geometry.attributes.position.count)
+		// console.groupEnd()
+
 		webglBuffer.numElements = geometry.attributes.position.count;
 
 		gl.bindVertexArray(webglBuffer.vao);
@@ -657,7 +664,7 @@ export class Renderer {
 				//attributeLocation = attributeLocations["aExtra"];
 			}else{
 				let attributeLocation = attributeLocations[attributeName].location;
-				
+
 				gl.vertexAttribPointer(attributeLocation, bufferAttribute.itemSize, type, normalized, 0, 0);
 				gl.enableVertexAttribArray(attributeLocation);
 			}
@@ -712,6 +719,11 @@ export class Renderer {
 		let worldView = new THREE.Matrix4();
 
 		let mat4holder = new Float32Array(16);
+
+		// console.group("Renderer.renderNodes")
+		// console.log("octree...", octree)
+		// console.log("nodes...", nodes)
+		// console.groupEnd()
 
 		let i = 0;
 		for (let node of nodes) {
@@ -920,9 +932,9 @@ export class Renderer {
 				let uFilterReturnNumberRange = material.uniforms.uFilterReturnNumberRange.value;
 				let uFilterNumberOfReturnsRange = material.uniforms.uFilterNumberOfReturnsRange.value;
 				let uFilterPointSourceIDClipRange = material.uniforms.uFilterPointSourceIDClipRange.value;
-				
-				
-				
+
+
+
 				shader.setUniform2f("uFilterReturnNumberRange", uFilterReturnNumberRange);
 				shader.setUniform2f("uFilterNumberOfReturnsRange", uFilterNumberOfReturnsRange);
 				shader.setUniform2f("uFilterPointSourceIDClipRange", uFilterPointSourceIDClipRange);
@@ -956,7 +968,7 @@ export class Renderer {
 				for(const attributeName in geometry.attributes){
 					const bufferAttribute = geometry.attributes[attributeName];
 					const vbo = webglBuffer.vbos.get(attributeName);
-					
+
 					gl.bindBuffer(gl.ARRAY_BUFFER, vbo.handle);
 					gl.disableVertexAttribArray(attributeLocation);
 				}
@@ -1003,7 +1015,7 @@ export class Renderer {
 					offset = Number.isNaN(offset) ? 0 : offset;
 
 					shader.setUniform1f("uExtraScale", scale);
-					shader.setUniform1f("uExtraOffset", offset);					
+					shader.setUniform1f("uExtraOffset", offset);
 				}
 
 			}else{
@@ -1018,11 +1030,11 @@ export class Renderer {
 
 						let type = this.glTypeMapping.get(bufferAttribute.array.constructor);
 						let normalized = bufferAttribute.normalized;
-						
+
 						gl.bindBuffer(gl.ARRAY_BUFFER, vbo.handle);
 						gl.vertexAttribPointer(attributeLocation, bufferAttribute.itemSize, type, normalized, 0, 0);
 						gl.enableVertexAttribArray(attributeLocation);
-						
+
 					}
 				}
 			}
@@ -1079,6 +1091,11 @@ export class Renderer {
 			}
 		}
 
+		// console.group("Renderer.renderOctree")
+		// console.log("octree...", octree)
+		// console.log("nodes...", nodes)
+		// console.groupEnd()
+
 		{ // UPDATE SHADER AND TEXTURES
 			if (!this.shaders.has(material)) {
 				let [vs, fs] = [material.vertexShader, material.fragmentShader];
@@ -1106,7 +1123,12 @@ export class Renderer {
 					`#define num_clippolygons ${numClipPolygons}`,
 				];
 
-
+				// debugger
+				// console.group("Renderer.renderOctree")
+				// console.log("octree...", octree)
+				// console.log("nodes...", nodes)
+				// console.log("octree.pcoGeometry.root.isLoaded()...", octree.pcoGeometry.root.isLoaded())
+				// console.groupEnd()
 				if(octree.pcoGeometry.root.isLoaded()){
 					let attributes = octree.pcoGeometry.root.geometry.attributes;
 
@@ -1215,7 +1237,7 @@ export class Renderer {
 			 }else{
 				 gl.depthMask(false);
 			 }
-			 
+
 		}
 
 
@@ -1233,10 +1255,10 @@ export class Renderer {
 			shader.setUniform1f("fov", Math.PI * camera.fov / 180);
 			shader.setUniform1f("near", camera.near);
 			shader.setUniform1f("far", camera.far);
-			
+
 			if(camera instanceof THREE.OrthographicCamera){
 				shader.setUniform("uUseOrthographicCamera", true);
-				shader.setUniform("uOrthoWidth", camera.right - camera.left); 
+				shader.setUniform("uOrthoWidth", camera.right - camera.left);
 				shader.setUniform("uOrthoHeight", camera.top - camera.bottom);
 			}else{
 				shader.setUniform("uUseOrthographicCamera", false);
@@ -1284,7 +1306,7 @@ export class Renderer {
 
 				const lClipSpheres = shader.uniformLocations["uClipSpheres[0]"];
 				gl.uniformMatrix4fv(lClipSpheres, false, flattenedMatrices);
-				
+
 				//const lClipSpheres = shader.uniformLocations["uClipSpheres[0]"];
 				//gl.uniformMatrix4fv(lClipSpheres, false, material.uniforms.clipSpheres.value);
 			}
@@ -1310,14 +1332,14 @@ export class Renderer {
 
 
 			shader.setUniform3f("uIntensity_gbc", [
-				material.intensityGamma, 
-				material.intensityBrightness, 
+				material.intensityGamma,
+				material.intensityBrightness,
 				material.intensityContrast
 			]);
 
 			shader.setUniform3f("uRGB_gbc", [
-				material.rgbGamma, 
-				material.rgbBrightness, 
+				material.rgbGamma,
+				material.rgbBrightness,
 				material.rgbContrast
 			]);
 
@@ -1443,7 +1465,7 @@ export class Renderer {
 
 		const gl = this.gl;
 
-		// PREPARE 
+		// PREPARE
 		if (target != null) {
 			this.threeRenderer.setRenderTarget(target);
 		}
